Moves Facebook Post Client (doing I/O and depending on framework classes) to infrastructure
