+++ src/test/kotlin/dev/pankowski/garcon/domain/LunchServiceTest.kt
@@ -1,5 +1,6 @@
 package dev.pankowski.garcon.domain
 
+import dev.pankowski.garcon.infrastructure.JsoupFacebookPostClient
 import dev.pankowski.garcon.infrastructure.persistence.InMemorySynchronizedPostRepository
 import dev.pankowski.garcon.infrastructure.persistence.someFailedRepost
 import dev.pankowski.garcon.infrastructure.persistence.someSuccessRepost
@@ -20,7 +21,7 @@ class LunchServiceTest : FreeSpec({
     // given
     val pageConfig = somePageConfig()
 
-    val postClient = mockk<FacebookPostClient>()
+    val postClient = mockk<JsoupFacebookPostClient>()
     val repository = mockk<SynchronizedPostRepository>()
     val service = spyk(
       LunchService(someLunchConfig(pages = listOf(pageConfig)), mockk(), repository, postClient, mockk(), mockk())
@@ -44,7 +45,7 @@ class LunchServiceTest : FreeSpec({
     val lastSeenPublishedAt = now()
     val lastSeen = somePost(publishedAt = lastSeenPublishedAt)
 
-    val postClient = mockk<FacebookPostClient>()
+    val postClient = mockk<JsoupFacebookPostClient>()
     val postClassifier = mockk<LunchPostClassifier>()
     val reposter = mockk<SlackReposter>()
     val repository = mockk<SynchronizedPostRepository>()
@@ -84,7 +85,7 @@ class LunchServiceTest : FreeSpec({
     val post = somePost()
     val classification = Classification.LunchPost
 
-    val postClient = mockk<FacebookPostClient>()
+    val postClient = mockk<JsoupFacebookPostClient>()
     val postClassifier = mockk<LunchPostClassifier>()
     val reposter = mockk<SlackReposter>()
     val repository = spyk(InMemorySynchronizedPostRepository())
@@ -115,7 +116,7 @@ class LunchServiceTest : FreeSpec({
     val post = somePost()
     val classification = Classification.MissingKeywords
 
-    val postClient = mockk<FacebookPostClient>()
+    val postClient = mockk<JsoupFacebookPostClient>()
     val postClassifier = mockk<LunchPostClassifier>()
     val reposter = mockk<SlackReposter>()
     val repository = spyk(InMemorySynchronizedPostRepository())
